+++ conftest.py
@@ -2,6 +2,8 @@
 #
 # SPDX-License-Identifier: MIT
 
+import importlib
+
 from doctest import ELLIPSIS
 
 import pytest
@@ -39,6 +41,12 @@ try:
     import sphinx  # noqa: F401
 except ImportError:
     collect_ignore.extend(["docs"])
+
+for name in ["fastapi", "pyramid", "flask", "aiohttp", "starlette"]:
+    try:
+        importlib.import_module(name)
+    except ImportError:  # noqa: PERF203
+        collect_ignore.extend([f"tests/integrations/test_{name}.py"])
 
 
 @pytest.fixture(name="svc")