+++ docs/examples/fastapi/simple_fastapi_app.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import os
 
-from typing import AsyncGenerator
+from collections.abc import AsyncGenerator
 
 from fastapi import FastAPI