+++ docs/examples/starlette/health_check.py
@@ -11,7 +11,7 @@ async def healthy(request: Request) -> JSONResponse:
     Ping all external services.
     """
     ok: list[str] = []
-    failing: list[dict[str, str]] = []
+    failing: dict[str, str] = {}
     code = 200
 
     for svc in svcs.starlette.get_pings(request):
@@ -19,7 +19,7 @@ async def healthy(request: Request) -> JSONResponse:
             await svc.aping()
             ok.append(svc.name)
         except Exception as e:
-            failing.append({svc.name: repr(e)})
+            failing[svc.name] = repr(e)
             code = 500
 
     return JSONResponse(