+++ tests/integrations/test_fastapi.py
@@ -8,19 +8,15 @@ from contextlib import asynccontextmanager
 
 import pytest
 
+from fastapi import FastAPI
+from fastapi.testclient import TestClient
+
 import svcs
 
 from tests.helpers import CloseMe
 
 
-try:
-    from fastapi import FastAPI
-    from fastapi.testclient import TestClient
-except ImportError:
-    pytest.skip("FastAPI not installed", allow_module_level=True)
-
-
-@pytest.mark.asyncio()
+@pytest.mark.asyncio
 @pytest.mark.parametrize("yield_something", [True, False])
 @pytest.mark.parametrize("cm", [True, False])
 async def test_integration(yield_something, cm):