+++ tests/integrations/test_starlette.py
@@ -8,23 +8,19 @@ from contextlib import asynccontextmanager
 
 import pytest
 
+from starlette.applications import Starlette
+from starlette.middleware import Middleware
+from starlette.responses import JSONResponse
+from starlette.routing import Route
+from starlette.testclient import TestClient
+
 import svcs
 
 from tests.fake_factories import async_bool_cm_factory, async_int_factory
 from tests.helpers import CloseMe
 
 
-try:
-    from starlette.applications import Starlette
-    from starlette.middleware import Middleware
-    from starlette.responses import JSONResponse
-    from starlette.routing import Route
-    from starlette.testclient import TestClient
-except ImportError:
-    pytest.skip("Starlette not installed", allow_module_level=True)
-
-
-@pytest.mark.asyncio()
+@pytest.mark.asyncio
 @pytest.mark.parametrize("yield_something", [True, False])
 @pytest.mark.parametrize("cm", [True, False])
 async def test_integration(yield_something, cm):
@@ -110,7 +106,7 @@ async def healthy(request):
     )
 
 
-@pytest.mark.asyncio()
+@pytest.mark.asyncio
 async def test_get_pings(registry, container):
     """
     Our get_pings returns registered pings.