+++ tests/test_registry.py
@@ -76,7 +76,7 @@ class TestRegistry:
         with cm():
             ...
 
-    @pytest.mark.asyncio()
+    @pytest.mark.asyncio
     async def test_async_generators_become_context_managers(self, registry):
         """
         If a generator-based factory is passed, it's automatically wrapped with
@@ -204,7 +204,7 @@ class TestRegistry:
 
         orc.assert_called_once_with()
 
-    @pytest.mark.asyncio()
+    @pytest.mark.asyncio
     async def test_async_context_manager(self, close_me):
         """
         The registry is also an async context manager that acloses on exit.
@@ -226,7 +226,7 @@ class TestRegistry:
         not hasattr(contextlib, "aclosing"),
         reason="Hasn't contextlib.aclosing()",
     )
-    @pytest.mark.asyncio()
+    @pytest.mark.asyncio
     async def test_async_empty_close(self, registry):
         """
         Asynchronously closing an empty registry does nothing.
@@ -236,7 +236,7 @@ class TestRegistry:
         async with svcs.Registry():
             ...
 
-    @pytest.mark.asyncio()
+    @pytest.mark.asyncio
     @needs_working_async_mock
     async def test_aclose_mixed(self, registry):
         """
@@ -258,7 +258,7 @@ class TestRegistry:
 
         async_close.assert_awaited_once()
 
-    @pytest.mark.asyncio()
+    @pytest.mark.asyncio
     @needs_working_async_mock
     async def test_aclose_logs_failures(self, registry, caplog):
         """
@@ -287,15 +287,19 @@ class TestRegistry:
         assert Service in registry
         assert AnotherService not in registry
 
-    def test_iter(self, registry):
+    def test_iterate(self, registry):
         """
-        Iterating over a registry returns all registered services types.
+        It's possible to iterate over the registered services.
         """
         registry.register_factory(Service, Service)
         registry.register_factory(AnotherService, AnotherService)
-        registry.register_value(YetAnotherService, YetAnotherService)
 
-        assert {Service, AnotherService, YetAnotherService} == set(registry)
+        assert {
+            svcs.RegisteredService(Service, Service, False, True, None),
+            svcs.RegisteredService(
+                AnotherService, AnotherService, False, True, None
+            ),
+        } == {rs for rs in registry}  # noqa: C416 -- explicit on purpose
 
     def test_gc_warning(self, recwarn):
         """
@@ -434,17 +438,14 @@ class TestTakesContainer:
 
         assert svcs._core._takes_container(module.factory)
 
-    def test_catches_invalid_sigs(self):
+    def test_ignores_invalid_sigs(self):
         """
-        If the factory takes more than one parameter, raise an TypeError.
+        If the first parameter is anything but what we handle, ignore it.
         """
 
         def factory(foo, bar): ...
 
-        with pytest.raises(
-            TypeError, match="Factories must take 0 or 1 parameters."
-        ):
-            svcs._core._takes_container(factory)
+        assert not svcs._core._takes_container(factory)
 
     def test_call_works(self):
         """