+++ tests/typing/core.py
@@ -5,15 +5,10 @@
 import contextlib
 import sys
 
-from typing import AsyncGenerator, Generator, NewType, Protocol
+from collections.abc import AsyncGenerator, Generator
+from typing import NewType, Protocol
 
 import svcs
-
-
-if sys.version_info >= (3, 9):
-    from typing import Annotated
-else:
-    from typing_extensions import Annotated
 
 
 reg = svcs.Registry()
@@ -142,7 +137,7 @@ if sys.version_info >= (3, 10):
 
 
 # Multiple factories for same type:
-S1 = Annotated[str, "s1"]
+S1 = NewType("S1", str)
 S2 = NewType("S2", str)
 
 reg.register_value(S1, "foo")