+++ tests/typing/pyramid.py
@@ -4,7 +4,7 @@
 
 from __future__ import annotations
 
-from typing import Generator
+from collections.abc import Generator
 
 import pyramid